Fixed correlation ellipse.
Fixed a bug in the ellipse drawing.  The ellipse represents a contour of a bivariate Normal distribution with the same covariance matrix as the sample covariance matrix of the plotted principal components.  When the number of observations is <= 2, the ellipse will not display because the sample covariance matrix will be degenerate.  

There was a bug in the conditional to check the number of observations that caused the off-diagonals to be ignored.  This has been fixed.